Arranging components into folders
