+++ src/components/GitHub/GitHubProjects.vue
@@ -64,7 +64,7 @@
 </template>
 
 <script>
-  import { HTTP } from '../http-common.js';
+  import { HTTP } from '../../http-common.js';
 
   export default{
     data() {
