+++ src/components/Projects/AddProject.vue
@@ -32,7 +32,7 @@
 </template>
 
 <script>
-import {HTTP} from '../http-common.js';
+import {HTTP} from '../../http-common.js';
 
 export default {
   name: 'addProj',
