+++ src/components/Projects/EditProject.vue
@@ -30,8 +30,8 @@
 </template>
 
 <script>
-	import { EventBus } from '../event-bus.js';
-	import {HTTP} from '../http-common.js';
+	import { EventBus } from '../../event-bus.js';
+	import {HTTP} from '../../http-common.js';
 	export default{
 		name: 'editProj',
 		data(){
