+++ src/components/Projects/Project.vue
@@ -25,7 +25,7 @@
         <delete-project></delete-project>
       </div>
       <div class="col-md-2" align="center">
-        <router-link v-bind:to="'/project/'+project.id+'/releases'">
+        <router-link v-bind:to="'/projects/'+project.id+'/releases'">
           <button type="button" class="btn btn-info btn-md falko-button">
             Releases
           </button>
@@ -36,9 +36,9 @@
 </template>
 
 <script>
-import DeleteProject from '@/components/DeleteProject';
-import EditProject from '@/components/EditProject';
-import { HTTP } from '../http-common.js';
+import DeleteProject from '@/components/Projects/DeleteProject';
+import EditProject from '@/components/Projects/EditProject';
+import { HTTP } from '../../http-common.js';
 
 export default{
   name: 'Project',
